+++ pyfaaster/aws/__init__.py
@@ -1,6 +1,3 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2016-present, CloudZero, Inc. All rights reserved.
 # Licensed under the BSD-style license. See LICENSE file in the project root for full license information.
-
-import handlers_decorators as handlers_decorators_v2
-import handlers_decorators_deprecated as handlers_decorators
